// src/public-api/model/articleDto.ts
/**
 * MISCore Shop API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ArticleCategoryDto } from './articleCategoryDto';
import { ArticleTagDto } from './articleTagDto';


export interface ArticleDto { 
    id?: number;
    title?: string | null;
    summary?: string | null;
    content?: string | null;
    slug?: string | null;
    seoTitle?: string | null;
    seoDescription?: string | null;
    seoKeywords?: string | null;
    imageUrl?: string | null;
    articleType?: string | null;
    status?: string | null;
    isFeatured?: boolean;
    viewCount?: number;
    allowComments?: boolean;
    createdAt?: string;
    updatedAt?: string | null;
    publishedAt?: string | null;
    authorId?: string | null;
    authorName?: string | null;
    articleTags?: Array<ArticleTagDto> | null;
    articleCategories?: Array<ArticleCategoryDto> | null;
}